Found an erroro in the dac7512 send command (DAC7512_send) which took an 8 bit int instead of a 16 bit integer as an argument.

+++ include/spi/dac7512.h
@@ -17,7 +17,7 @@ __inline void DAC7512_unset_addr();
 void DAC7512_addr_init();
 void DAC7512_init(spi_device_t *dev);
 
-void DAC7512_send(spi_device_t *dev, uint16_t data, uint8_t addr);
+void DAC7512_send(spi_device_t *dev, uint16_t data, uint16_t addr);
 
 int16_t DAC7512_chip_select(uint16_t cs_arg);
